# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CleanCity Connect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type MeRole = (typeof MeRole)[keyof typeof MeRole];

export const MeRole = {
  user: "user",
  admin: "admin",
} as const;

export interface Me {
  id: number;
  displayName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: MeRole;
  points: number;
  totalReports: number;
  verifiedReports: number;
  pendingReports: number;
}

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface Report {
  id: number;
  imagePath: string;
  locationLabel: string;
  latitude: number;
  longitude: number;
  description: string;
  status: ReportStatus;
  pointsAwarded: number;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  /** @nullable */
  decidedAt?: string | null;
}

export type MapReportStatus =
  (typeof MapReportStatus)[keyof typeof MapReportStatus];

export const MapReportStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface MapReport {
  id: number;
  latitude: number;
  longitude: number;
  status: MapReportStatus;
  locationLabel: string;
  createdAt: string;
}

export type AdminReportStatus =
  (typeof AdminReportStatus)[keyof typeof AdminReportStatus];

export const AdminReportStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface AdminReport {
  id: number;
  imagePath: string;
  locationLabel: string;
  latitude: number;
  longitude: number;
  description: string;
  status: AdminReportStatus;
  pointsAwarded: number;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  /** @nullable */
  decidedAt?: string | null;
  reporterId: number;
  reporterName: string;
  /** @nullable */
  reporterEmail?: string | null;
}

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface AdminUser {
  id: number;
  displayName: string;
  /** @nullable */
  email?: string | null;
  role: AdminUserRole;
  points: number;
  totalReports: number;
  verifiedReports: number;
  createdAt: string;
}

export interface CreateReportBody {
  imagePath: string;
  /** @minLength 1 */
  locationLabel: string;
  latitude: number;
  longitude: number;
  /** @minLength 1 */
  description: string;
}

export type DecideReportBodyDecision =
  (typeof DecideReportBodyDecision)[keyof typeof DecideReportBodyDecision];

export const DecideReportBodyDecision = {
  verified: "verified",
  rejected: "rejected",
} as const;

export interface DecideReportBody {
  decision: DecideReportBodyDecision;
  adminNote?: string;
  pointsAwarded?: number;
}

export type SetRoleBodyRole =
  (typeof SetRoleBodyRole)[keyof typeof SetRoleBodyRole];

export const SetRoleBodyRole = {
  user: "user",
  admin: "admin",
} as const;

export interface SetRoleBody {
  role: SetRoleBodyRole;
}

export interface Reward {
  id: number;
  name: string;
  description: string;
  pointsCost: number;
  category: string;
  available: boolean;
}

export interface Redemption {
  id: number;
  rewardId: number;
  rewardName: string;
  pointsSpent: number;
  code: string;
  createdAt: string;
}

export interface RedeemBody {
  rewardId: number;
}

export interface OverviewStats {
  totalReports: number;
  verifiedReports: number;
  pendingReports: number;
  totalPointsAwarded: number;
  contributors: number;
}

export interface TopReporter {
  id: number;
  displayName: string;
  points: number;
  verifiedReports: number;
}

export interface AdminStats {
  pendingCount: number;
  verifiedToday: number;
  rejectedToday: number;
  totalUsers: number;
  topReporters: TopReporter[];
}

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Forbidden
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Bad request
 */
export type BadRequestResponse = ErrorResponse;

export type AdminListReportsParams = {
  status?: AdminListReportsStatus;
};

export type AdminListReportsStatus =
  (typeof AdminListReportsStatus)[keyof typeof AdminListReportsStatus];

export const AdminListReportsStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;