# lib/api-zod/src/generated/types/decideReportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CleanCity Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { DecideReportBodyDecision } from "./decideReportBodyDecision";

export interface DecideReportBody {
  decision: DecideReportBodyDecision;
  adminNote?: string;
  pointsAwarded?: number;
}