# lib/api-zod/src/generated/types/mapReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CleanCity Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { MapReportStatus } from "./mapReportStatus";

export interface MapReport {
  id: number;
  latitude: number;
  longitude: number;
  status: MapReportStatus;
  locationLabel: string;
  createdAt: Date;
}