# lib/api-zod/src/generated/types/redemption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CleanCity Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface Redemption {
  id: number;
  rewardId: number;
  rewardName: string;
  pointsSpent: number;
  code: string;
  createdAt: Date;
}